requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-generativeai==0.4.1
gunicorn==21.2.0
